### 2024-09-20 #7252 @editorialbot
[PRE REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis
**Submitting author:** @jcross186 (<a href="http://orcid.org/0000-0002-2338-3644">James Ross</a>)
**Repository:** https://github.com/acil-bwh/bayes_traj/
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.1
