### 2024-09-23 #7252 @Kevin-Mattheus-Moerman
- [ ] The paper YAML is not formatted as per our standards. Please review and edit the paper accordingly: https://joss.readthedocs.io/en/latest/example_paper.html
