### 2024-09-26 db783a1
Adding contribution guidelines and linking from README
