+++ CONTRIBUTING.md
@@ -1,0 +1,8 @@
+# Contributing to *bayes_traj*
+
+Contributions to *bayes_traj* are encouraged and valued. Please
+first read the available 
+[Documentation](https://acil-bwh.github.io/bayes_traj/index.html).
+If you have a usage question, bug report, or feature enhancement
+request, please contact the lead developer,
+[James C. Ross](mailto:jcross186@gmail.com).
