+++ paper.bib
@@ -6,7 +6,8 @@
   number={3},
   pages={347},
   year={2022},
-  publisher={American Psychological Association}
+  publisher={American Psychological Association},
+  doi={10.1037/met0000359}
 }
 
 @article{van2020overview,
@@ -16,7 +17,8 @@
   volume={43},
   pages={100323},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.alcr.2019.100323}
 }
 
 
@@ -88,14 +90,17 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
   doi     = {10.1038/s41592-019-0686-2},
 }
 
-@incollection{ross2022dysanapsis,
-  title={Dysanapsis Is Associated with COPD Trajectories At-Risk for Mortality and Exacerbations},
-  author={Ross, J and San Jose Estepar, R and Ash, S and Pistenmaa, CL and Han, MK and Bhatt, SP and Bodduluri, S and Charbonnier, J-P and Washko, GR and Diaz, AA},
-  booktitle={C18. COPD: WHO, WHY, AND HOW},
-  pages={A3725--A3725},
-  year={2022},
-  publisher={American Thoracic Society}
+
+@article{ross2024dysanapsis,
+  title={Dysanapsis is differentially related to lung function trajectories with distinct structural and functional patterns in COPD and variable risk for adverse outcomes},
+  author={Ross, James C and Est{\'e}par, Raul San Jos{\'e} and Ash, Sam and Pistenmaa, Carrie and Han, MeiLan and Bhatt, Surya P and Bodduluri, Sandeep and Sparrow, David and Charbonnier, Jean-Paul and Washko, George R and others},
+  journal={EClinicalMedicine},
+  volume={68},
+  year={2024},
+  publisher={Elsevier},
+  doi={10.1016/j.eclinm.2023.102408}
 }
+
 
 @article{ross2018longitudinal,
   title={Longitudinal modeling of lung function trajectories in smokers with and without chronic obstructive pulmonary disease},
@@ -117,7 +122,8 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
   number={1},
   pages={343--354},
   year={2016},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TMI.2016.2608782}
 }
 
 @article{lange2015lung,
@@ -128,7 +134,8 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
   number={2},
   pages={111--122},
   year={2015},
-  publisher={Mass Medical Soc}
+  publisher={Mass Medical Soc},
+  doi={10.1056/NEJMoa1411532}
 }
 
 @article{agusti2019lung,
