### 2024-09-26 #7252 @jcross186
Hi Kevin,

I have checked in changes that address previous issues. Please let me know if all looks good.

Kind regards,

-James

From: Kevin Mattheus Moerman ***@***.***>
Date: Thursday, September 26, 2024 at 11:52 AM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7252)

        External Email - Use Caution

@jcross186<https://secure-web.cisco.com/1c6ujiesvXndssHe4z9iLsbGuCI4ww5aGc3UNTKyy4_wJ4iduKJT9fpFwMIkoDw5_zkxdQJ0_EVT9PCnLcOL7UbCoN1C_V26-QJCL2sfWIfD7H5YxYRrPaZ8uQ5LbMlSHe1LLhhUk51y87rfCfCkDFp0_H3Iv3I1KLlVqgs-BF9XEtR-Zc7gTT8-5sobL4mK3trfDGMv1EXN898j0nxFVb6aIXxylRa0G8aP6LiFFSfAB02t8Tq59Zn6EyPC811_xiUhdNaxxyk14w-2i-oVIwnJztyBMDfB-ZWdZBMs9e4FLOSor8qZIUftIWobj_8CK/https%3A%2F%2Fgithub.com%2Fjcross186> 👋
