### 2024-09-27 #7252 @majensen
Hi @jcross186 - if you have time, can you browse https://reviewers.joss.theoj.org/lookup for reviewers that you think might be appropriate? You can put usernames without the @ sign in this thread. This can help speed up the review process - much appreciated!
