### 2024-09-27 #7252 @jcross186
Hi Mark,

Here are a few:

gvieralopez
limengbinggz
mrazomej
matteodelucchi
gchure
nhtran93
lindemann09

Kind regards,

-James

From: Mark Jensen ***@***.***>
Date: Friday, September 27, 2024 at 2:31 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7252)

        External Email - Use Caution

Hi @jcross186<https://secure-web.cisco.com/1SB-a5bWxLXgWVKhRmRAiH585-ohye_sUTeW81fMSeFg_zgIa5kHHPZPYGgTdhqhfWKoYlDg3lqIP_YzH8uRnLj_1WJonPYSf_OukSQI6v5wJ5pEm-lCwKQOheV-I5nwar4RU1VCu7RWC5KSvpMjP0iKKE277EPYVT6A9l7K2HJl6iamgw_NN_moQKwlgekF88w9eAn73imKADAPyUYnjH2m4_xjVDdzU2Ck6pPyXCksN7uQKqYkiIbxB4krua6Us8JSODOUI90ZtMWE4bLbAFrNl_RhV-Rsb7GxMNkA7-uQPaj_FSz6K7r8VWvnApOla/https%3A%2F%2Fgithub.com%2Fjcross186> - if you have time, can you browse https://reviewers.joss.theoj.org/lookup<https://secure-web.cisco.com/1Y1u-L3XUkfHRhUjjmwxWNYr-2RyyqFAKrVuN2cJEJpamL-Y1iS_LwRCwWhs3XQ_wnEl2Ykx58GM4PzNPrmnK5VskGFLrbjMp8RlSWBsEGndwJhl8YYKwhCZ1OY6zGjgFjYHOLxJMAmQPCBT9GfGAgJDqd8mAzJLS6MXzRFJIs-b8GwTIdVvE7meY45T-FP4JUCoMtHSb9J5k8xhb5wijE1I_FYmxoYqg07T9ZpMAM0zG_d1GmN04VrUS64em21KRzdPD8Ebq-Jm5j0wovt4Y17saFQDr7uJe3LCvWXFa5QxtDrrC6w_S2nb6xrYgo4AZ/https%3A%2F%2Freviewers.joss.theoj.org%2Flookup> for reviewers that you think might be appropriate? You can put usernames without the @ sign in this thread. This can help speed up the review process - much appreciated!
