### 2024-10-05 #7252 @majensen
Hi @mrazomej - would you consider reviewing [this work](#7252) for [JOSS](https://joss.theoj.org)? We would appreciate your expertise. Thanks --@majensen
