### 2024-10-06 #7252 @gvieralopez
> Hi @gvieralopez - would you be willing to review [this work](#7252) for [JOSS](https://joss.theoj.org)? We could use your expertise. Thanks for considering! --@majensen

Yes
