### 2024-10-13 5797cca
Minor change to ensure backwards compatibility
