+++ bayes_traj/generate_prior.py
@@ -757,6 +757,8 @@ def main():
         with open(op.model, 'rb') as f:
             print("Reading model...")
             mm = pickle.load(f)['MultDPRegression']
+            if 'ranef_indices_' not in dir(mm):
+                mm.ranef_indices_ = None
             if op.model_trajs is not None:
                 pg.set_model(mm, \
                     np.array(op.model_trajs.split(','), dtype='int'))
