### 2024-10-13 43ec3b2
Adding waic2_thresh as input argument
