### 2024-10-15 99b31b9
Can now specify how many samples to use when computing WAIC2
