+++ bayes_traj/summarize_traj_model.py
@@ -24,6 +24,8 @@ def main():
     parser.add_argument('--hide_ic', help='Use this flag to hide compuation \
         and display of information criterai (BIC and WAIC2), which can take \
         several moments to compute.', action="store_true")
+    parser.add_argument('-s', help='Number of samples to use when computing \
+        WAIC2', type=int, default=100)
     
     op = parser.parse_args()
     
@@ -47,7 +49,9 @@ def main():
         waic2 = None
     else:
         bic = mm.bic()
-        waic2 = mm.compute_waic2()
+        assert isinstance(op.s, int), 'Number of samples must be an integer.'
+        assert op.s > 0, 'Number of samples must be greater than 0.'
+        waic2 = mm.compute_waic2(op.s)
     
     # Compute fit stats
     ave_pps = ave_pp(mm)
