### 2024-10-27 #7323 @gvieralopez
@majensen

I find the work to be very promising, with the manuscript itself being the primary area for improvement. With some revisions, I believe the work will meet the standards expected for JOSS. Below are some minor comments that may help the author enhance the quality of the submission, particularly addressing items currently marked as incomplete in my review:

1. **License File**: The `license.txt` file in the repository still contains placeholder text from the MIT license template: “Copyright (c) [year] [fullname].”

2. **Authorship**: The paper currently lists only one author. However, after reviewing the repository's commit history, it appears that another contributor has made significant contributions: https://github.com/fritzo.

3. **Substantial Scholarly Effort**: Initially, I had reservations about whether the work met the journal's requirement for "substantial scholarly effort." However, the software documentation clarifies the significant contributions and potential impact of this work. This said, the manuscript itself would benefit from a thorough revision to clearly communicate the library’s scope, its positioning within the current landscape of open-source solutions, its limitations, and its value to a general audience. Although JOSS has strict length limitations, this work has much to offer, and it would be unfortunate to publish it in its current form.

4. **State of the Field**: The manuscript’s review of the field is quite limited, citing only two papers, both over two years old. Recent work in trajectory analysis, including some contributions published within JOSS, could be referenced to give readers a better sense of the existing landscape. Addressing “what is currently available” and explaining “how this contribution builds upon or complements existing tools” is essential to provide context for this work.

5. **Community Guidelines**: The documentation would benefit from the addition of community guidelines for external contributors.
