### 2024-10-30 47d3f17
Modifying how parameters are initialized when NaN (specifically, w_mu_0)
