### 2024-11-04 #7323 @jcross186
@majensen<https://secure-web.cisco.com/169VNUGO64Yaw5dcMgTu4LkWq4JvoTU-e-2wJoNV6TMOfXZojlMxPwhEkWxp4PHLff5R4U6GSg83-NSGtJit0Uf8ejIYKJqoAjQzx0Ye5zPJBoPSoEgISksV3_dgGgJnTySZxJ3D7Os1FOrIRP95Y3glDB1qHqb3A7OJIOa22uNtT6b9qX2aSU8CQGu35y48AFoQq5lDJxyRAP1Xk34g9PDcUFSwZ10w8uaNMIvi7W8iQ0WVcA4Ov65rdZoKXIQvMGhIjYuM_6aCVXV4VsoVigOv44Qof0koaRNgDRZKr-guq87FrZCxW0FCDDW0aOUyv/https%3A%2F%2Fgithub.com%2Fmajensen>

Hi Mark,

Just confirming that we’re still waiting for one more review?

Many thanks!

-James

From: Mark Jensen ***@***.***>
Date: Saturday, October 26, 2024 at 8:57 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

@lindemann09<https://secure-web.cisco.com/1KxrAOQ9RNb3NSm7YktN0bR7gJRI5Mgqki8eRAkaS8D3oay-liix9v98L6MWQYjEmcqff6FHlxF6kvFaPwKlMU2xjMcQibkQ1M9lxms0CCBOKhsO6Mz5PXyAPVaFA8DCuc_S4DqdHa4RAm1JJfz9UZBSk_Cm0W8AwDMXk4gWQHgP58mZudgbHo5PJzp17_Cr1qjyUsfV9ojIvs95_b_wXS6l1fs9p5gN-YmPnJub6SGnPhtfB6L4Z8j56pLXnUkqIb9KBYdbqNpKNiKo5SKlF8QqUVr92CCGMAbeuGiNUH4xCPULR9RbRjKNL31k4s-ql/https%3A%2F%2Fgithub.com%2Flindemann09> @gvieralopez<https://secure-web.cisco.com/18xhz9u8LYn1GshmpbwL5IW8AlqHfBqXyAfsWF99YOYT0N0ORHsoiZRjH7tqltYTMxE7g3u3aAd7BR_jDxlhH8rEI1ZuNtbsVQiQkrTxT-SEnVQM0KMlgm0ii4jxLTcOjfmOek8v1-IZYICbJI1xySVqKMbulDCmJAQ5JIpN0UrxDIlhqJ9D3v7XjvGqIRNUoPlMIL5vQmJfyCHzsh5wYf0MJ-WH__5s05NJUQcCbMa3TcRcGjUWCQ4rNOI_1TVPneJ0awn4ZOz31_4n3ERLRo_ohbN49JA_O7G6S69myJ-8yo8wh9aIx0TpjGdHt2-tC/https%3A%2F%2Fgithub.com%2Fgvieralopez> looking forward to your reviews of #7323<https://secure-web.cisco.com/1gnxTcBilHteBf-ibRVxQMZKmV7stidvoYti51n5WT89DNg7_moAQ4bmHjyr1fFZ3xuHUIFs1DkvPmmfr8P6ayk9NFbTrEbdvoxTnnOQZNsO1Dk4x0kKfU6j5LZYbwd60lwqOsM7e7r5w3ScwppjBUt66-jTY150TWq1NHMKEosFnXAn78F6W4yDx_kBTnw4_r4fwnf6CbKvGazFMGjumXoKmU-2lDZy2da2UNxFfO8Z6Prb8yfBj0YskxCAwWID0xCvju_I5UOfa-CpCjWAkZTmRXEnb7PUkAXufOYcJvXiWKxhg6LrRNAoqs9qyIkjT/https%3A%2F%2Fgithub.com%2Fopenjournals%2Fjoss-reviews%2Fissues%2F7323> - thanks!
