### 2024-11-22 #7323 @jcross186
Hi Mark,

Just circling back. Looking forward to addressing reviewer comments!

Kind regards,

-James

From: Mark Jensen ***@***.***>
Date: Monday, November 4, 2024 at 5:24 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

@jcross186<https://secure-web.cisco.com/1J-yzKc8KkF8Qf8jT8bG58Ji4eMvvqT5peb8NkdPr6zXXzpX9izhjtcQoW_5wdt-yRlnaK1PjyDipdPUwLMrPoLciWJP1TmOSlqucRI2CPWePpVD_qG8fY6r6r2GsvDm9q3gf2SMX_JKCHcbGOluoCcWZTohdNriOXuvBqJsOZP_zz3Pr5t8NmYNnqr23OA4_Goafcp8oH-pO_IaaWyUeovuMDNTr6THIVy-UIkKl5N5YIS9h-lWbvJUaiGAv3CBM_wWveKA0OWTfG_YmwNWzmLg7Er-4aTwfGPC7GDlj1SVWP2Daq2LObS77yRwyaDo5/https%3A%2F%2Fgithub.com%2Fjcross186> - yes, hoping that @lindemann09<https://secure-web.cisco.com/1_rlA0XVGqzMGB1tGvbvY13lIGqmCt8HRMPZFkvWIDVVSiBqmSpCs4irDP7DxqXxeLK4GetO8oZK8GT2uEsdmv93oSPfFBSEmZ_UAkZFB21-NVR3frnNMNIVUSCsfVppo6NJmTyDY2BQ1cJPYHLG3GDXb2s3lRdd4Gl8gfp8LTQ6ynpgPRuJxg8cHcMoideCm1Vc6FICdwNhoDvm20SbThjbAUCaRT1S7VYbAN7HbOyNAAVoEm-B70FO4BFZ8JyYtP3XGNB2q4Cc_v4PCXQF9h8res032hIzS_28vVo8iloj-FWGXsNJAGAWLAsdK2uEr/https%3A%2F%2Fgithub.com%2Flindemann09> will update us. I will look for alternate reviewers too. Thanks for the ping!
