### 2024-11-23 #7323 @gvieralopez
> @gvieralopez have you had a chance to review the areas that remain on your checklist? @lindemann09 looking forward to your review. Thanks! --@majensen

@majensen I am waiting for the author reply to my comments related to those areas before considering those clear. No further review from my side on this round.

Best regards!
