### 2024-12-18 bdf2de2
added more detailed community guidelines
