+++ CONTRIBUTING.md
@@ -1,8 +1,32 @@
 # Contributing to *bayes_traj*
 
-Contributions to *bayes_traj* are encouraged and valued. Please
-first read the available 
-[Documentation](https://acil-bwh.github.io/bayes_traj/index.html).
-If you have a usage question, bug report, or feature enhancement
-request, please contact the lead developer,
-[James C. Ross](mailto:jcross186@gmail.com).
+Thank you for your interest in contributing to **bayes_traj**! Contributions
+are encouraged and highly valued, whether they involve reporting issues,
+suggesting enhancements, or submitting code improvements.
+
+## How to Contribute
+1. **Read the Documentation**: Before contributing, please familiarize yourself with the [documentation](https://acil-bwh.github.io/bayes_traj/index.html).
+2. **Fork the Repository**: Create a fork of the repository on GitHub.
+3. **Create a Branch**: Work on your contribution in a separate branch based on the `main` branch.
+4. **Submit a Pull Request**:
+   - Provide a clear description of your contribution.
+   - Reference any related issues or feature requests.
+   - Ensure your changes are well-documented and include relevant tests.
+
+For detailed steps, refer to [GitHub’s guide on contributing](https://docs.github.com/en/get-started/quickstart/contributing-to-projects).
+
+## How to Report Issues
+If you encounter a bug or have a feature request, please:
+1. Check if the issue has already been reported in the [Issues section](https://github.com/acil-bwh/bayes_traj/issues).
+2. If not, create a new issue with:
+   - A clear title.
+   - Steps to reproduce the problem.
+   - Relevant system information (e.g., Python version, dependencies).
+   - Any error messages or screenshots.
+
+## How to Seek Support
+For usage questions or assistance, you can:
+1. Review the available [documentation](https://acil-bwh.github.io/bayes_traj/index.html).
+2. Contact the lead developer, [James C. Ross](mailto:jcross186@gmail.com).
+
+Thank you for helping make **bayes_traj** better!
