### 2024-12-18 7831976
Paper revamp based on JOSS reviewer feedback
