+++ paper.md
@@ -19,35 +19,57 @@ bibliography: paper.bib
 
 # Statement of need 
 
-Trajectory analysis broadly refers to the application of methods to
-longitudinal data to identify distinct patterns of change and assign study
-subjects to their most likely trajectory group.
+Trajectory analysis broadly refers to techniques and modeling paradigms that
+explain heterogeneity in longitudinal data. These methods identify
+the most suitable number of subgroups (trajectories) in the data, the distinct
+patterns of change characterizing each trajectory, and the most likely assignment of
+study participants to trajectories. Methods of trajectory analysis have
+been applied to a wide range of fields including pyschology, criminology,
+behavioral research, and epidemiology. (These methods are distinct from those that
+track people, animials, vehicles,
+and natural phenomena -- also referred to as trajectory analysis -- and which
+have their own dedicated set of techniques and frameworks. See, e.g., 
+[@shenk2021traja] and [@viera2023pactus]).
+
 Although trajectory analysis has been applied in mutiple
 domains, the motivation for developing **bayes_traj** has been to improve our
 understanding of
 heterogeneity in the context of chronic obstructive pulmonary disease (COPD), a
 leading cause of death worldwide. Research has shown that there are multiple
 patterns of lung function development and decline, with some patterns associated
-with greater risk of developing COPD [@lange2015lung]. Researchers have applied techniques
+with greater risk of developing COPD [@lange2015lung]. Furthermore, there
+is a growing recognition that COPD is better conceived of as a multi-faceted syndrome,
+requiring consideration of other disease facets (such as clinical presentation
+and structural assesment from medical images) [@lowe2019copdgene].
+Researchers have applied techniques
 of trajectory analysis to longitudinal measures of lung function to delineate
 distinct patterns of progression for further analysis [@agusti2019lung].
 Existing trajectory approaches
 are predominantly frequentist in nature and use maximum likelihood to identify point
-estimates of unknown parameters. These approaches do not permit incorporation of prior informaiton.
+estimates of unknown parameters. These approaches do not permit incorporation of
+prior informaiton.
 Challenges arise when study cohorts lack sufficient longitudinal data characteristics
-to adequately power frequentist-based trajectory algorithms. Furthermore, there
-is a growing recognition that COPD is better conceived of as a multi-faceted syndrome,
-requiring consideration of other disease facets (such as clinical presentation
-and structural assesment from medical images) [@lowe2019copdgene]. This
-motivates development and
-application of scalable approaches that can simultaneously model distinct progression
-patterns across multiple health measures, especially in data-limited scenarios.
+to adequately power frequentist-based trajectory algorithms.
+Bayesian approaches are well-suited for data-limited scenarios given their ability
+to incorporate prior knowledge in the model fitting process, though existing
+Bayesian trajectory approaches
+use sampling-based inference (i.e. Markov chain Monte Carlo)
+which can be slow to converge and can suffer from the so-called "lable-switching" problem.
+There is thus a need for scalable approaches that can simultaneously model distinct
+progression patterns across multiple health measures, especially in data-limited scenarios.
 
 # Summary
 
 **bayes_traj** is a Python package developed to perform Bayesian trajectory analysis.
-Although our primary motivation is to improve understanding of COPD heterogeneity,
-the package makes no domain-specific assumptions and is generally applicable.
+While its development was initially motivated by the need to better understand
+chronic obstructive pulmonary disease (COPD) heterogeneity, the package is
+domain-agnostic and can be applied to a wide range of disciplines.
+The package is designed for researchers who need scalable methods for trajectory
[…]
