+++ paper.bib
@@ -1,3 +1,80 @@
+@misc{tan2022bcclong,
+  title={BCClong package: Bayesian consensus clustering model for mixed-type longitudinal data},
+  author={Tan, Z and Shen, C and Lu, Z},
+  year={2022}
+}
+
+
+@article{komarek2014capabilities,
+  title={Capabilities of R package mixAK for clustering based on multivariate continuous and discrete longitudinal data},
+  author={Kom{\'a}rek, Arno{\v{s}}t and Kom{\'a}rkov{\'a}, Lenka},
+  journal={Journal of Statistical Software},
+  volume={59},
+  pages={1--38},
+  year={2014}
+}
+
+
+@article{komarek2013clustering,
+  title={Clustering for multivariate continuous and discrete longitudinal data},
+  author={Kom{\'a}rek, Arno{\v{s}}t and Kom{\'a}rkov{\'a}, Lenka},
+  year={2013}
+}
+
+
+@article{lock2013bayesian,
+  title={Bayesian consensus clustering},
+  author={Lock, Eric F and Dunson, David B},
+  journal={Bioinformatics},
+  volume={29},
+  number={20},
+  pages={2610--2616},
+  year={2013},
+  publisher={Oxford University Press}
+}
+
+
+@article{lu2024clustering,
+  title={Clustering Longitudinal Data: A Review of Methods and Software Packages},
+  author={Lu, Zihang},
+  journal={International Statistical Review},
+  year={2024},
+  publisher={Wiley Online Library}
+}
+
+
+@article{lu2023joint,
+  title={Joint clustering multiple longitudinal features: A comparison of methods and software packages with practical guidance},
+  author={Lu, Zihang and Ahmadiankalati, Mojtaba and Tan, Zhiwen},
+  journal={Statistics in Medicine},
+  volume={42},
+  number={29},
+  pages={5513--5540},
+  year={2023},
+  publisher={Wiley Online Library}
+}
+
+
+@article{viera2023pactus,
+  title={pactus: A Python framework for trajectory classification},
+  author={Viera-L{\'o}pez, Gustavo and Morgado-Vega, JJ and Reyes, Alfredo and Altshuler, Ernesto and Almeida-Cruz, Yudivi{\'a}n and Manganini, Giorgio},
+  journal={Journal of Open Source Software},
+  volume={8},
+  number={89},
+  pages={5738},
+  year={2023}
+}
+
+@article{shenk2021traja,
+  title={Traja: A Python toolbox for animal trajectory analysis},
+  author={Shenk, Justin and Byttner, Wolf and Nambusubramaniyan, Saranraj and Zoeller, Alexander},
+  journal={Journal of Open Source Software},
+  volume={6},
+  number={63},
+  pages={3202},
+  year={2021}
+}
+
 @article{zang2022bayesian,
   title={Bayesian estimation and model selection in group-based trajectory models.},
   author={Zang, Emma and Max, Justin T},
