### 2024-12-18 #7323 @jcross186
@gvieralopez
Thank you very much for your thoughtful review of the manuscript and your helpful comments. Below is a point-by-point response to your comments:

1. License file:
Placeholders have been replaced with appropriate values.

2. Authorship:
Thank you for your thoughtful observation regarding the authorship of this paper. While Fritz Obermeyer (fritzo) contributed code to the project as part of a subcontract funded through my NIH R01 grant, his contributions, though valuable, were more limited in scope and were made specifically to incorporate Pyro probabilistic programming language functionality within bayes_traj. However, this functionality continues to be under development, and none of bayes_traj’s current capabilities relies on pyro or these software contributions.

Given the comparatively limited scope of Fritz's involvement in these aspects for this project, I determined that authorship was not appropriate. Instead, I have added an acknowledgment of Dr. Obermeyer’s contributions, which I believe strikes an appropriate balance in recognizing his contributions without overstating the role played in the project (as either first or last author).

3. Substantial scholarly effort:
The manuscript has been expanded to clarify the library’s scope, its position within the current open-source landscape, its limitations, and its value to the general audience.

4. State of the field:
The description of the state of the field has been expanded. Given the breadth of available modeling assumptions and implementations, as well as space limitations, the reader is directed to several recent and comprehensive reviews. Additionally, two broad categories (GBTM and LCMM) are described with specific implementations referenced to contextualize the contributions of bayes_traj.

5. Community guidelines:
I have updated the CONTRIBUTING.md file to include more detailed community guidelines. These now provide explicit instructions for contributing code, reporting issues, and seeking support, in line with JOSS criteria.

Many thanks again for committing time to review this manuscript! Please let me know if there are additional items you believe could further strengthen this contribution!
