### 2024-12-18 #7323 @gvieralopez
I am satisfied with the revisions made to your manuscript. The quality of
the work has significantly improved, and I am happy to recommend acceptance
for publication.
I did notice a few minor typos:
* "animals" in the first paragraph of the statement of need
* "label switching"
Please double-check for any additional errors before final submission.

Thank you for your hard work and dedication to this open source project.

Tomorrow morning, I will check all the boxes on the review checklist
because right now I don't have access to my GitHub account (I am responding
through Gmail).

Best regards,

Gustavo
