### 2024-12-19 #7323 @jcross186
Many thanks again for your review! I have just committed the typo corrections.

Kind regards,

-James

From: Gustavo Viera López ***@***.***>
Date: Wednesday, December 18, 2024 at 4:27 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution
I am satisfied with the revisions made to your manuscript. The quality of
the work has significantly improved, and I am happy to recommend acceptance
for publication.
I did notice a few minor typos:
* "animals" in the first paragraph of the statement of need
* "label switching"
Please double-check for any additional errors before final submission.

Thank you for your hard work and dedication to this open source project.

Tomorrow morning, I will check all the boxes on the review checklist
because right now I don't have access to my GitHub account (I am responding
through Gmail).

Best regards,

Gustavo
