### 2025-01-31 #7323 @jcross186
Greetings,

Any update on the review?

-James

From: Mark Jensen ***@***.***>
Date: Thursday, December 19, 2024 at 1:48 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

@lindemann09<https://secure-web.cisco.com/1ptpXShqB4tsNp1g8-FZfb_ePT2CIMlfHPTJaFWy6CnGnr9Fs1JCBoO7REAKaG30ej3YeCeWPVSxqkW2EaI4m-0UnKd1QNHgFz56mbWThmAnRzb3qVq6KO3ivBuGkBLHcqzSTqcg0HLsbqYLlwJUY-LD1kqMKmbkZbar7QIn-BZAfFvTtm241866C1p415nIc4nVkvCya6eRsyL3FvN9WbiWHLsj2XH75yLMhTrQtboY6W2cGlYm8H47VeQI99sVx9rx8oJKReIAqkBme7vADN1R-8NHX3BZHoCsdBy8tmF7OS_3NwZJ8n1k4NgWxrr0n/https%3A%2F%2Fgithub.com%2Flindemann09> Are you able yo get us your review soon? Thanks.
