### 2025-02-07 #7323 @jcross186
Hi Mark,

Should we consider an additional reviewer?

Best,

-James

From: Mark Jensen ***@***.***>
Date: Saturday, February 1, 2025 at 1:34 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

@lindemann09<https://secure-web.cisco.com/1SSEhCkwXTcYI56MLFiKDgW0wCiILbNPdzzlz_leItBsNMuT2rab6SVhi5zAUG9Mf7fZNM8puFonx1BIjPTUbwo0Xln0eDVXsztH7NZV9mMxQuDx-7d67aqi_JRhR8ZIiX7ZOB_UNuweAu0EvLPL_QJWeB7rFR8fuXYD6mChOaSw_NrdVGATYiKOtAr777AYrbQrZp4iZnYA3It1Ry0g0wGQ5slN1Ldd2gRxEUxK8rqtY2slY4fnvFke4DN1T11hYK-Qz-HOLjSjJpRHnsmds6bCYfUfBhnR4XxpbjVTz6u_-3F6OqWlPchHh1beci7tr/https%3A%2F%2Fgithub.com%2Flindemann09>, can you let us know in the next couple of days whether you're able to complete your review? No hard feelings if not, but then I'll be able to explore other avenues for accelerating the process. I really appreciate this. Mark
