### 2025-02-10 #7323 @jcross186
Thanks very much, Mark!

-James

From: Mark Jensen ***@***.***>
Date: Monday, February 10, 2025 at 12:48 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

Hi @jcross186<https://secure-web.cisco.com/1APGVqBkGhmc3wv0KUCNwu8EXRAO9A0Cm5B15vVf7Lmt4_-noW34GLtcdZvxj1SCRtI3KuzlQ41EHkxeRgatftfOwrsJcS5OEOh_5yQuKu4c4gFJju5ggNgZCqjuQEgOxAGGXFJOtHf5dQw_ytjBMZ3gY-QSRr5JtwusVoaR_mi1oOflmtEtCmll2ubclp9ZT04WwWUaZKukT6wkwDTDdzIJGmvm_fZu55K9IsfAb3Tvq5aUTtQKNromNQXWDiudhCoDfYVTjn-50NAU4OdbaBmNIUXedIzdDzYG6ndOkO9PSwwliT3Ms_hGYgo-01xfP/https%3A%2F%2Fgithub.com%2Fjcross186> - yes, unfortunately that looks like our best way forward. I will pursue this.
