### 2025-02-17 65f6f21
Now getting version from __init__.py
