+++ bayes_traj/__init__.py
@@ -1,0 +1,13 @@
+# Define version
+__version__ = "1.0.2"
+
+# Import key modules to make them accessible
+from . import bayes_traj_main
+from . import viz_data_prior_draws
+from . import viz_model_trajs
+from . import generate_generic_data
+from . import summarize_traj_model
+from . import assign_trajectory
+from . import update_model
+from . import get_alpha_estimate
+from . import generate_prior
