### 2025-02-24 acil-bwh/bayes_traj#4 @gchure
JOSS Review: Paper Comments and Authorship
Hi! I'm reviewing your paper for [JOSS](https://github.com/openjournals/joss-reviews/issues/7323#issuecomment-2677094547) and am opening issues as I progress through my checklist. This pertains to comments on the manuscript and a question about authorship.

## Paper Comments

The paper is well written overall. I think details about COPD are not exactly necessary for the readership, but I am fine with keeping that detail in. My comments are minor and are as follows:

1. The formatting of lines 43 - 53 is broken as the bullet points show up as a list of asterisks.

2. I think it would be useful to specify `bayes_traj` is a command line tool and include some information about its API. I very short figure showing a brief example of how to specify and evaluate the model would really make the paper stand out.

## Authorship

I noted that you acknowledged Tingting Zhao for aiding with model equations and Fritz Obermeyer for contribution of code. Both seem important for the software, and Fritz Obermeyer is listed as a contributor on the repository. Should they be included as authors on this work?
