### 2025-02-28 ebeebe6
Updated formatting to ensure proper markdown formatting for list of features
