### 2025-02-28 6f979f6
Incorporating workflow figure and caption
