### 2025-02-28 acil-bwh/bayes_traj#4 @jcross186
Hi Griffin,

Thank you very much for taking the time to review this paper. Responses to your comments below.

**1. The formatting of lines 43 - 53 is broken as the bullet points show up as a list of asterisks.**
Thank you for pointing this out. I have updated the formatting to correct this.

**2. I think it would be useful to specify bayes_traj is a command line tool and include some information about its API. A very short figure showing a brief example of how to specify and evaluate the model would really make the paper stand out.**
Thank you for your suggestion. I revised the Summary section to explicitly state that bayes_traj provides a suite of command-line tools for prior specification, model fitting, and posterior evaluation. I included a figure illustrating the key command-line tools in bayes_traj, their relationships, and their role in the modeling workflow. The caption includes a high-level description of each command-line tool, outlining its function and relationship to other tools in the workflow.

**3. Authorship: I noted that you acknowledged Tingting Zhao for aiding with model equations and Fritz Obermeyer for contribution of code. Both seem important for the software, and Fritz Obermeyer is listed as a contributor on the repository. Should they be included as authors on this work?**
Thank you for your thoughtful observation regarding the authorship of this paper. While Fritz Obermeyer (fritzo) contributed code to the project as part of a subcontract funded through my NIH R01 grant, his contributions, though valuable, were more limited in scope and were made specifically to incorporate Pyro probabilistic programming language functionality within bayes_traj. However, this functionality continues to be under development, and none of bayes_traj’s current capabilities relies on pyro or these software contributions. Given the comparatively limited scope of Fritz's involvement in these aspects for this project, I determined that authorship was not appropriate. Instead, I have added an acknowledgment of Dr. Obermeyer’s contributions, which I believe strikes an appropriate balance in recognizing his contributions without overstating the role played in the project (as either first or last author). Likewise, although Dr. Zhao helped derive variational update equations for inference over binary target variables, she did not contribute code to the bayes_traj package. (We are pursuing a separate manuscript focused on the theoretical aspects of this contribution). For this reason, I decided that recognizing her in the acknowledgements was most appropriate.
