### 2025-02-28 acil-bwh/bayes_traj#4 @jcross186
Hi Griffin,

Thanks very much for your review! Responses provided though GitHub.

Kind regards,

-James

From: Griffin Chure ***@***.***>
Date: Monday, February 24, 2025 at 1:38 AM
To: acil-bwh/bayes_traj ***@***.***>
Cc: Subscribed ***@***.***>
Subject: [acil-bwh/bayes_traj] JOSS Review: Paper Comments and Authorship (Issue #4)

        External Email - Use Caution

Hi! I'm reviewing your paper for JOSS<https://secure-web.cisco.com/1MGDHTboXRQ7Etf-an2aPgNQ1si9hIJe-wMPDaHQcw12SMPrtIwRAqMNyHTE4NzDts-fhdD3663BOPnl-ImPNhB8Atd2X-X98FwYkp2taowwJKLShst1GyWwSZ9jHVq2vcQoX4d_zqY2FzSEUIKQc-aw8ECzVlDRbvvwBVmxNQmPN40RHvawiFz6Mv6ojz0sqDDXVhehE3M4TXKZVjpBXSpk7xVYBqiZjfvHjWzsp3v2C-PeK5SGaxl2IKqEqQ32GX6oRIDV7eDUqRTuS9ft7keaDkQIvYtHF_y0MwiPju6Hde4MCMEaexGmvbPTEVEqR/https%3A%2F%2Fgithub.com%2Fopenjournals%2Fjoss-reviews%2Fissues%2F7323%23issuecomment-2677094547> and am opening issues as I progress through my checklist. This pertains to comments on the manuscript and a question about authorship.

Paper Comments

The paper is well written overall. I think details about COPD are not exactly necessary for the readership, but I am fine with keeping that detail in. My comments are minor and are as follows:

  1.  The formatting of lines 43 - 53 is broken as the bullet points show up as a list of asterisks.
  2.  I think it would be useful to specify bayes_traj is a command line tool and include some information about its API. I very short figure showing a brief example of how to specify and evaluate the model would really make the paper stand out.

Authorship

I noted that you acknowledged Tingting Zhao for aiding with model equations and Fritz Obermeyer for contribution of code. Both seem important for the software, and Fritz Obermeyer is listed as a contributor on the repository. Should they be included as authors on this work?
