### 2025-03-06 acil-bwh/bayes_traj#4 @gchure
Hi @jcross186, thanks for the detailed response! I follow your argument for not including Dr. Fritz Obermeyer as an author. For Dr. Tingting Zhao, however, I think that authorship may still be deserved. While Dr. Zhao may not have directly contributed code, their contribution of the variational update equations feels critical to the operation and utility of `bayes_traj`. I do not believe that their inclusion as an author on this work would preclude their authorship of a more detailed theoretical study. If it is true that Dr. Zhao's mathematical contribution was necessary for the release of `bayes_traj`, I strongly believe they should be listed as an author, unless they are opposed.
