### 2025-03-08 acil-bwh/bayes_traj#4 @jcross186
@gchure -- Thank you for your thoughtful feedback. bayes_traj represents the continuation of my PhD work and is the product of a multi-year NIH career development award. As such, it reflects a substantial investment of my personal time, effort, and expertise. Given this background, the authorship situation is somewhat unique in that I am the sole developer, implementer, and author of the software and manuscript.
Dr. Zhao’s contribution—deriving the variational update equations for binary target variables—was the result of a project in which I mentored her. While this was a valuable mathematical addition to the methodology, it represents a targeted contribution to one aspect of the software, rather than a foundational or integrative role in its development.
First authorship typically implies leadership in conceptualization, software design, code development, algorithm implementation, performance evaluation, manuscript writing, figure generation, and reviewer responses, none of which fell under Dr. Zhao’s role in this project.
Last authorship is generally reserved for high-level supervision, mentorship, or funding, which Dr. Zhao did not provide in this case.
Dr. Zhao’s contribution, while important in its own right, was limited in scope and was one of several components of bayes_traj. Given the unique circumstances of this one-author work, I believe the most appropriate way to recognize her efforts is through an acknowledgment in the Acknowledgments section.
I appreciate the opportunity to clarify this point and value your perspective in ensuring proper attribution of contributions.

From: Griffin Chure ***@***.***>
Date: Thursday, March 6, 2025 at 2:40 PM
To: acil-bwh/bayes_traj ***@***.***>
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [acil-bwh/bayes_traj] JOSS Review: Paper Comments and Authorship (Issue #4)

        External Email - Use Caution

Hi @jcross186<https://secure-web.cisco.com/1OKA0t3lbHBFdwfGtnOauNzwFF7nu60dWyt7I4qo_zWbCp5vtKZPt5_6ywziTMdsX_pLgT4z2u4rvBdPq-d-ArEuQFiwENwZ-0Ck0Xb8ELzw9ZynRUo6RS-twiWzmAuAstMMom_4kY9wVCUikkqphWv84AbwWfrBsmUg7zv29FrW4gn6PcR4XFIDDUAatHkyw3oTN1yhTT4mspITIuwlvG2kuVv7KbyC-8C5T5WeLQmz2j2P0cc_Bs4cb0ixevUn8goNiWoutF2xMyUhY4fV_LrEqcw9mF7oNNmSuASca4B521qdCOsbeZOM-9JriaFSKhSADvu3wJXJ9W_I_PxZhLQ/https%3A%2F%2Fgithub.com%2Fjcross186>, thanks for the detailed response! I follow your argument for not including Dr. Fritz Obermeyer as an author. For Dr. Tingting Zhao, however, I think that authorship may still be deserved. While Dr. Zhao may not have directly contributed code, their contribution of the variational update equations feels critical to the operation and utility of bayes_traj. I do not believe that their inclusion as an author on this work would preclude their authorship of a more detailed theoretical study. If it is true that Dr. Zhao's mathematical contribution was necessary for the release of bayes_traj, I strongly believe they should be listed as an author, unless they are opposed.
