### 2025-03-10 acil-bwh/bayes_traj#4 @gchure
Hi @jcross186, thanks for the response. I do follow your point, but I do maintain that Dr. Zhao should be an author on the work. Developing the equations is important to the work, as you note. In my measure, it does not matter if their contribution applies to only one portion of the software as that feature is being shipped as part of `bayes_traj` in general.

While I understand the implications of first and last authorship, they are not sufficient to justify leaving Dr. Zhao as solely an acknowledgement. I don't think there's any reason you cannot be first author and corresponding author. In that case, if Dr. Zhao is technically the "last" author, the correspondence can still be directed to you. This approach has worked well for me, where my [recent paper in JOSS](https://joss.theoj.org/papers/10.21105/joss.06270) has a "last" author who did not contribute code to the project but assisted in it none the less. In that case, I was noted as the first and corresponding author. As a final point, I personally believe in being generous with authorship -- my philosophy is that if there is any question whether someone contributed or "made it possible", they should be listed as an author to be safe. I know this is not the JOSS policy, however, so I think we should get another opinion.

To move forward, I would really appreciate the viewpoint of the editor @majensen and Dr. Zhao. If either or both of them think leaving Dr. Zhao as an acknowledgment is sufficient, I'm happy to consider this issue closed.
