### 2025-03-10 acil-bwh/bayes_traj#4 @jcross186
Hi @gchure – I appreciate the thoughtful exchange. I have updated the author list as suggested.

From: Griffin Chure ***@***.***>
Date: Monday, March 10, 2025 at 12:20 AM
To: acil-bwh/bayes_traj ***@***.***>
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [acil-bwh/bayes_traj] JOSS Review: Paper Comments and Authorship (Issue #4)

        External Email - Use Caution

Hi @jcross186<https://secure-web.cisco.com/1Sc1vUJzbbPf7WmGXqBngRoTBYA2IuVVK0Nmp_nuY65YnJ_6YcmvxjksARr0-GGIKQ4GDb8x13vmq-O6TqKTfjq13qWw4slaBA1-FC_we2ZhZPWQXICS3t7eQO9drGRE_pC3b0R7LQI7Lh6owG7xl89EWMraBCKYdVeZvmk01aMWT1BObp0ETwAMSe7bwhDp7N4O8ja8RpFK13haF5Sz9QvKKCxlDYxsuAxnQGcYpWhXak16D5x2Jt84eINRGQXWZxdk8cQ0aNZDTxvS5OrURknSiAuW8CV-yD0wKeIfrkzEodkb_xwvWXUOIAMPssEe4/https%3A%2F%2Fgithub.com%2Fjcross186>, thanks for the response. I do follow your point, but I do maintain that Dr. Zhao should be an author on the work. Developing the equations is important to the work, as you note. In my measure, it does not matter if their contribution applies to only one portion of the software as that feature is being shipped as part of bayes_traj in general.

While I understand the implications of first and last authorship, they are not sufficient to justify leaving Dr. Zhao as solely an acknowledgement. I don't think there's any reason you cannot be first author and corresponding author. In that case, if Dr. Zhao is technically the "last" author, the correspondence can still be directed to you. This approach has worked well for me, where my recent paper in JOSS<https://secure-web.cisco.com/1eW4W1mU8T0-TpliapOuYaoVTa5GykNOPAglsW4d3QikN5fdF6Hz9SMqPfACM_LuLV3LYWL--iveVYgKNgosn1TYtx_djxEt-jscEUTrHuEMYcrzmY6jbaYmqNQumHAy1NRsqeWCCCvoc-Z2wHk8B19AAh-ygnQsa8fDhdvVZFI-xWmZhOz7O7iANJSa5zx_G5X-4fHCYaQHdtNDQqQ1DBE1eqDN8b1Y_ShSgy1Z6BC7wFcsqYjbZbgBrmGj8kw3pJcSR42otsoLJG0VuNGhKNFyu5JsM-qf4lDhS-r8HtAVciHjU5ZqRIR66Y6z8g0xB/https%3A%2F%2Fjoss.theoj.org%2Fpapers%2F10.21105%2Fjoss.06270> has a "last" author who did not contribute code to the project but assisted in it none the less. In that case, I was noted as the first and corresponding author. As a final point, I personally believe in being generous with authorship -- my philosophy is that if there is any question whether someone contributed or "made it possible", they should be listed as an author to be safe. I know this is not the JOSS policy, however, so I think we should get another opinion.

To move forward, I would really appreciate the viewpoint of the editor @majensen<https://secure-web.cisco.com/1drlfDl5JOaOy8WjF54facfBqMXbrgy3o6dxRNNVQpMAXC9E8dzf78mm5HYaqZciDmuXixgjEvBl3RNxWvzJYOtjdcApBKH1YTm8MtrlTw6nT0aalU_ges-QynaZwcfd8jtRfwu4j_ZmWVPlLTxjqc3hG5Lh4UCdDntrWqwbTyq143FxD6eVb5h_90NfzIBhYQSi5Q6dz3JDR2N3vXnycrxmLW-teWaxezQMoN-f99kPtVwkodlBndFyGH_4WLJK8OYmSFrgH9DB27vIh6he-0S9eeZTuXpRDh5yQ7yPpCdo-VyT5KTHdbPiJiZwx3M0r/https%3A%2F%2Fgithub.com%2Fmajensen> and Dr. Zhao. If either or both of them think leaving Dr. Zhao as an acknowledgment is sufficient, I'm happy to consider this issue closed.
