### 2025-03-12 acil-bwh/bayes_traj#5 @gchure
JOSS Review: More Detail in Contribution Guidelines
Hi! I'm reviewing your paper for [JOSS](https://github.com/openjournals/joss-reviews/issues/7323#issuecomment-2677094547) and am opening issues as I progress through my checklist. This pertains to the contribution guidelines.

You mention in the [contribution guidelines](https://github.com/acil-bwh/bayes_traj/blob/master/CONTRIBUTING.md) that questions or problems with the software should be addressed to @jcross186 via email. A major component of open-source software is the public discussion of problems/features/bugs via an issue tracker (like this one!). Do you have any preference for how issues should be structured? If you do not want to use issues, could you please explain why?
