### 2025-03-12 acil-bwh/bayes_traj#4 @gchure
Great, thank you @jcross186. I'm satisfied with this issue and have now closed the issue.
