### 2025-03-12 acil-bwh/bayes_traj#6 @gchure
JOSS Review: Failing Test and Automated Testing
Hi! I'm reviewing your paper for [JOSS](https://github.com/openjournals/joss-reviews/issues/7323#issuecomment-2677094547) and am opening issues as I progress through my checklist. This pertains to tests.

### Failing test
Cloning the repository and running `pytest` within the directory has one failing test that I haven't been able to debug. The failing test is `test_compute_prior_info_5` which has an assertion error and the log is provided at the end of this issue. Do you have an idea why this is failing, and is it working on your end?

### Setting Up Automated Testing

It would be great if `bayes_traj` would have a CI/CD pipeline set up where the tests are run upon push to main (or whatever your release branch may be). This should be doable using standard [GitHub Actions](https://resources.github.com/learn/pathways/automation/essentials/application-testing-with-github-actions/).

```python
platform darwin -- Python 3.10.9, pytest-8.2.2, pluggy-1.5.0
rootdir: /Users/gchure/Desktop/bayes_traj
configfile: pyproject.toml
plugins: anyio-3.5.0, html-4.1.1, cov-4.1.0, metadata-3.0.0, codecov-0.5.1
collected 24 items

bayes_traj/tests/test_fit_stats.py .                                     [  4%]
bayes_traj/tests/test_generate_prior.py ........F                        [ 41%]
bayes_traj/tests/test_mult_dp_regression.py ............                 [ 91%]
bayes_traj/tests/test_utils.py ..                                        [100%]

=================================== FAILURES ===================================
__________________________ test_compute_prior_info_5 ___________________________

    def test_compute_prior_info_5():
        """Model and data, different preds, new target
        """
        model_file_name = os.path.split(os.path.realpath(__file__))[0] + \
            '/../resources/models/model_1.p'
        data_file_name = os.path.split(os.path.realpath(__file__))[0] + \
            '/../resources/data/trajectory_data_1.csv'

        mm = pickle.load(open(model_file_name, 'rb'))['MultDPRegression']
        mm.cast_to_torch()
        mm.ranef_indices_ = None

        df = pd.read_csv(data_file_name)
        df['y2'] = df.y.values + 10

        targets = ['y', 'y2']
        preds = ['intercept', 'age', 'age^2']
        pg = PriorGenerator(targets, preds)
        pg.set_model(mm)
        pg.set_data(df, 'id')
        pg.compute_prior_info()

        # The following values are from a prior that -- upon visual inspection of
        # random draws -- appears reasonable. This test, then, should be considered
        # a regression test as opposed to a test of "correct" values
        prior_info_gt = \
            {'w_mu0': {'y': {'intercept': 8.102185344934247,
                             'age': -1.7021817743184755,
                             'age^2': 0.018416334689522876},
                       'y2': {'intercept': 18.10218534493426,
                              'age': -1.7021817743184746,
                              'age^2': 0.01841633468952318}},
             'w_var0': {'y': {'intercept': 1.9028074002722193,
                              'age': 0.1804899197433961,
                              'age^2': 0.0007043047299188618},
                        'y2': {'intercept': 1.9028074002722182,
                               'age': 0.18048991974339593,
                               'age^2': 0.0007043047299188608}},
             'lambda_a0': {'y': 14.399999999999999, 'y2': 14.399999999999999},
             'lambda_b0': {'y': 355.56217578065207, 'y2': 355.56217578065207},
             'w_mu': {'intercept': {'y': np.array([5.73256565, 9.27107768]),
                                    'y2': np.array([15.73256565, 19.27107768])},
                      'age': {'y': np.array([-1.91955117, -0.86877812]),
                              'y2': np.array([-1.91955117, -0.86877812,])},
                      'age^2': {'y': np.array([-0.0042961, -0.00525871,]),
                                'y2': np.array([-0.0042961, -0.00525871,])}},
             'w_var': {'intercept': {'y': np.array([ 0.06041792, 0.30262284,]),
                                     'y2': np.array([0.06041792, 0.30262284,])},
                       'age': {'y': np.array([0.00333069, 0.01713044]),
                               'y2': np.array([0.00333069, 0.01713044])},
                       'age^2': {'y': np.array([9.82818683e-06, 5.18071486e-05]),
                                 'y2': np.array([9.82818683e-06, 5.18071486e-05])}},
             'lambda_a': {'y': np.array([203.50423622, 11.95284849]),
                          'y2': np.array([203.50423622, 11.95284849])},
             'lambda_b': {'y': np.array([201.74450982, 48.8934525]),
                          'y2': np.array([201.74450982, 48.8934525])},
             'v_a': np.array([51., 51.]),
             #'v_b': np.array([50.74102343, 0.74102343]),
             'v_b': np.array([0.74102343, 50.74102343]),
             'traj_probs': np.array([0.5, 0.5]),
             'alpha': 0.5}

>       assert pg.prior_info_['w_mu0']['y']['intercept'] == \
            prior_info_gt['w_mu0']['y']['intercept'], "Error in prior"
E       AssertionError: Error in prior
E       assert 8.10218534493425 == 8.102185344934247

bayes_traj/tests/test_generate_prior.py:456: AssertionError
=============================== warnings summary ===============================
bayes_traj/tests/test_mult_dp_regression.py::test_update_u
bayes_traj/tests/test_mult_dp_regression.py::test_update_u
bayes_traj/tests/test_mult_dp_regression.py::test_update_u
bayes_traj/tests/test_mult_dp_regression.py::test_update_u
  /Users/gchure/Desktop/bayes_traj/bayes_traj/mult_dp_regression.py:1024: FutureWarning: DataFrameGroupBy.grouper is deprecated and will be removed in a future version of pandas.
    groupby_col = self.gb_.grouper.names[0]

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
=========================== short test summary info ============================
FAILED bayes_traj/tests/test_generate_prior.py::test_compute_prior_info_5 - A...
=================== 1 failed, 23 passed, 4 warnings in 2.53s ===================

```
