### 2025-03-12 #7323 @gchure
Hi @majensen, I've now finished off my review checklist. I've opened two issues on the `bayes_traj` main repository pertaining to testing and contribution guidelines. Once @jcross186 addresses these, I'll be able to finish off my review completely.
