+++ bayes_traj/tests/test_generate_prior.py
@@ -452,7 +452,7 @@ def test_compute_prior_info_5():
          'v_b': np.array([0.74102343, 50.74102343]),         
          'traj_probs': np.array([0.5, 0.5]),
          'alpha': 0.5}
-
+    pdb.set_trace()
     assert pg.prior_info_['w_mu0']['y']['intercept'] == \
         prior_info_gt['w_mu0']['y']['intercept'], "Error in prior"
     assert pg.prior_info_['w_mu0']['y']['age'] == \
