### 2025-03-13 c086c34
Updating test to reflect updated resource model
