+++ bayes_traj/tests/test_generate_prior.py
@@ -453,30 +453,30 @@ def test_compute_prior_info_5():
          'traj_probs': np.array([0.5, 0.5]),
          'alpha': 0.5}
 
-    assert pg.prior_info_['w_mu0']['y']['intercept'] == \
-        prior_info_gt['w_mu0']['y']['intercept'], "Error in prior"
-    assert pg.prior_info_['w_mu0']['y']['age'] == \
-        prior_info_gt['w_mu0']['y']['age'], "Error in prior"
-    assert pg.prior_info_['w_mu0']['y']['age^2'] == \
-        prior_info_gt['w_mu0']['y']['age^2'], "Error in prior"
-    assert pg.prior_info_['w_mu0']['y2']['intercept'] == \
-        prior_info_gt['w_mu0']['y2']['intercept'], "Error in prior"
-    assert pg.prior_info_['w_mu0']['y2']['age'] == \
-        prior_info_gt['w_mu0']['y2']['age'], "Error in prior"
-    assert pg.prior_info_['w_mu0']['y2']['age^2'] == \
-        prior_info_gt['w_mu0']['y2']['age^2'], "Error in prior"
-    assert pg.prior_info_['w_var0']['y']['intercept'] == \
-        prior_info_gt['w_var0']['y']['intercept'], "Error in prior"
-    assert pg.prior_info_['w_var0']['y']['age'] == \
-        prior_info_gt['w_var0']['y']['age'], "Error in prior"
-    assert pg.prior_info_['w_var0']['y']['age^2'] == \
-        prior_info_gt['w_var0']['y']['age^2'], "Error in prior"
-    assert pg.prior_info_['w_var0']['y2']['intercept'] == \
-        prior_info_gt['w_var0']['y2']['intercept'], "Error in prior"
-    assert pg.prior_info_['w_var0']['y2']['age'] == \
-        prior_info_gt['w_var0']['y2']['age'], "Error in prior"
-    assert pg.prior_info_['w_var0']['y2']['age^2'] == \
-        prior_info_gt['w_var0']['y2']['age^2'], "Error in prior"
+    assert np.isclose(pg.prior_info_['w_mu0']['y']['intercept'], \
+        prior_info_gt['w_mu0']['y']['intercept']), "Error in prior"
+    assert np.isclose(pg.prior_info_['w_mu0']['y']['age'], \
+        prior_info_gt['w_mu0']['y']['age']), "Error in prior"
+    assert np.isclose(pg.prior_info_['w_mu0']['y']['age^2'], \
+        prior_info_gt['w_mu0']['y']['age^2']), "Error in prior"
+    assert np.isclose(pg.prior_info_['w_mu0']['y2']['intercept'], \
+        prior_info_gt['w_mu0']['y2']['intercept']), "Error in prior"
+    assert np.isclose(pg.prior_info_['w_mu0']['y2']['age'], \
+        prior_info_gt['w_mu0']['y2']['age']), "Error in prior"
+    assert np.isclose(pg.prior_info_['w_mu0']['y2']['age^2'], \
+        prior_info_gt['w_mu0']['y2']['age^2']), "Error in prior"
+    assert np.isclose(pg.prior_info_['w_var0']['y']['intercept'], \
+        prior_info_gt['w_var0']['y']['intercept']), "Error in prior"
+    assert np.isclose(pg.prior_info_['w_var0']['y']['age'], \
+        prior_info_gt['w_var0']['y']['age']), "Error in prior"
+    assert np.isclose(pg.prior_info_['w_var0']['y']['age^2'], \
+        prior_info_gt['w_var0']['y']['age^2']), "Error in prior"
[…]
