### 2025-03-13 4a1ea82
Updating license and contribution guidelines
