### 2025-03-13 acil-bwh/bayes_traj#5 @jcross186
@gchure -- thanks for pointing this out! In fact, the contribution guidelines were updated in the 'joss' branch but not merged into 'master'. I have just merged the latest version of CONTRIBUTIONS.md. Thanks again!
