### 2025-03-13 acil-bwh/bayes_traj#6 @jcross186
@gchure -- that's for reporting this! I recently updated the resource model that is compared against in this test. I have updated the assertions to use 'isclose' to reflect tolerable discrepancies. The test was failing on my end as well but now passes. I'm closing this issue, but please let me know if you continue to see issues!
