### 2025-03-13 #7323 @jcross186
@gchure – thanks again for your thoughtful review of this paper. I have responded to both issues. Please let me know if you have additional comments, questions, or suggestions!

Kind regards,

-James

From: Griffin Chure ***@***.***>
Date: Wednesday, March 12, 2025 at 1:55 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

Hi @majensen<https://secure-web.cisco.com/11yNyaDprYuoPa-Cdl7s3TBRC77GS-kVfOSAt0kkNxotu3F8ye3HTJeLhGgW24DOt_HP3ImAudQF8144nhBz6q8lw0vai9AvtOJxbRfLTn6NakZaoVMqEEbtKbPukgST25G0HelpNAUV4eXU9P4vj2z-rZXGGAjgU89NhedUxpoEiillKlUcfAZqRZl2nXom-Uk8sNz7i4ntlzKu25bVVD7dlPBNpif-ZmppwuVWtx2ohsZX9fmCxtgt2HNx6JYyAB_eusQz_WCsTTWP2ACAu8DEbo9KjDwDYJ4f8DS59Z1zd5vBze8snF5DNpbwRWnmDCqV27WjSkDVfC_iZqG3FUg/https%3A%2F%2Fgithub.com%2Fmajensen>, I've now finished off my review checklist. I've opened two issues on the bayes_traj main repository pertaining to testing and contribution guidelines. Once @jcross186<https://secure-web.cisco.com/1zdnHG7khiB1X5dKIV7z_6QIzjpaox8ii0xOUXT588dacR2gAWtOCqZ5EGwf0H8v5snKgp_lJvZE0btrNuk_Egv5dszIiVYbul76MLPCTu1IqzJ_udiekJ3UkXSZ6pG1dcjSqd3jzWEUnw8K0S3nJKqLNx6lf9HFheOcPraoA6VDoMJOEqBr7X8Awblo3UdrwFSoFdPZdaPNVjevP5N_retyrzGZ0w-fFI0Ie2ObvyatpVePgIN4L6DTWqnfUQIcpu43CYJboaxqkcnsM9xkOaFHqI5EDt-aqgVCTYNsICvtH1vtBiDpUV-YNonqHauag3XetPpASV6RlzELrVfpuOQ/https%3A%2F%2Fgithub.com%2Fjcross186> addresses these, I'll be able to finish off my review completely.
