### 2025-03-14 f9a73e4
Removing version 3.8 as it is not supported
