+++ .github/workflows/pytest.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.10"]
+        python-version: ["3.8", "3.9", "3.10"]
 
     steps:
     - name: Checkout repository
@@ -22,10 +22,14 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
 
-    - name: Install dependencies
+    - name: Install package dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -e .  # Install package and dependencies from setup.py
+        pip install -e .  # Install the package
+
+    - name: Install testing dependencies
+      run: |
+        pip install pytest pytest-cov  # Only install testing-related dependencies
 
     - name: Run tests with coverage
       run: |
