### 2025-03-14 #7323 @jcross186
Hi @gchure – I have add pytest.yml to the master branch.  I’m seeing that the action triggers on github and the current suite of tests are passing. Thanks for the suggestion!

From: Griffin Chure ***@***.***>
Date: Thursday, March 13, 2025 at 2:56 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

Hi @jcross186<https://secure-web.cisco.com/1MYp8phK_RoGYoWzfwg74o52lgb605lGWJaSEx2Y6Ge9KGJCYznrN85VwlDLTGhtOxns474FXFPVfgTSafJ3R-e-pyVLFgQBC8C99aGhbM1aLAC6OyH342GoTZOnum7b4G0H-x0is1-lrLPA42VTAya74tmRzdKgMPdbhb-qpYOVQhQHmYi_W9UGqw40f0Ni_HHUPF8kCTystwml__7MQ33687jn-Rn3ibVBmTdPRJAWH3f_1gHVDj6zELLzonEDQYZR9dRNvEF0UIFGrXU8ioycnRm39FLcsVRJJPRpyhifLq3Tx2PzpJeLir7p09a6T/https%3A%2F%2Fgithub.com%2Fjcross186> , thanks for responding to the issues! I left a comment on the closed issue about the testing. It would be great if you could implement CI/CD with the test suite you have. Once that's done, I'm happy to check off everything on my review.
