### 2025-03-15 #7323 @gchure
I've now checked off the final box in my checklist -- I think `bayes_traj` is good to go for publication! Thanks @majensen for editing and @jcross186 for their rapid responses to all of my comments.
