### 2025-03-15 #7323 @majensen
@gchure - thanks so much for your thorough evaluation and the quick turnaround!
@jcross186 - I will do a quick readthrough as editor which may generate some minor fixes. One thing we will need is a public archive of the codebase, using Zenodo, FigShare, or similar. A short tutorial is [here](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729). Please let me know the version number of the archive and the DOI that you receive. Thanks!
