### 2025-03-17 #7323 @jcross186
Sounds good – should I wait for your read-through before uploading to Zenodo / DOI retrieval?

From: Mark Jensen ***@***.***>
Date: Friday, March 14, 2025 at 10:24 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

@gchure<https://secure-web.cisco.com/1YEglaj82Udwhqb-yyt6CGzqDYwRxz4t5RhKm97kS7DdPKpJiQXwhESMeElxJu3wl4F-RwYzr6V9Bw-G-LAbxjDApUER583QpXA8gtbU4YoFSWMy-pOstXttKQ-FTsA3mqpDNmJoCgc89Oc7UNJgmmBFYtgx0KV88z-fIm4MXVW1xp1XVsWLE-29Tx4ji1znU_ghW2S0AucJJUX8GiRCpV9BdgouDpUU2Hm38oSyCxxZxEA8nM-ZqkxlZc8s-mh4lHlzEVtb9euZSqFC7R3XqIdyQ14NmZwERKx4UJPytPJ20XFVKAFRyoZZkRGUx_rYW/https%3A%2F%2Fgithub.com%2Fgchure> - thanks so much for your thorough evaluation and the quick turnaround!
@jcross186<https://secure-web.cisco.com/17qnqLQNAQwsV_E6YgEQ4K1odtGOQbrGdgacd-QQWHSjUnpjyBf82uQSP8HM1YDyggqPYP8pK5JpNvsTwAznRzJNhTTG86Zre9AcetuHgR6LLH_Z93qkKLAlYVNIsJJ_BUOR9lihHlC_m1jIJwYNjqhsmK07EmRxsiqCVuS77nAbKosSAyU2Isgn659ug9AnV0gtZynrWdyR-d5O4Bha8n-xL5yNvITLQ7KdAQ5mfGu7nF0vO7GCgIlpxjefFLhaNtg-OUzob07piTa1hlPnGyeSbokdqiYZLbf_sMRWeYpYgbkEP6X2weWvJYLARIXOC/https%3A%2F%2Fgithub.com%2Fjcross186> - I will do a quick readthrough as editor which may generate some minor fixes. One thing we will need is a public archive of the codebase, using Zenodo, FigShare, or similar. A short tutorial is here<https://secure-web.cisco.com/1t6-7pLh11upLGUC6x6cC4tscNXjMsBXO9e129srGncweSD929DpKDlNSqssNDF45JOHe10mwzwZyTyu2sxREDrcyR3hzUTXf0ok_DbuyGGC8qXtjDit33mg3_4fVbm9n48_Plam2yMhfeXuOiMLATMXCfwGdsOSJYoQMLG6GgTfRQzvpm1iHdjdqrKuLaIv5K76gjEj-mSaGznFcrKXwB_k4AOGanl1xoko9Ux_aPHvmEWaSCxcrmA0KkBUFz8Q-zXxb6EV8gLPx0FAsz0XjNUb3DiQ-n3G6vo701zZsqPihfTWUbSijOnKvXh6zA6LA/https%3A%2F%2Fgithub.com%2Fopenjournals%2Fjoss-reviews%2Fissues%2F1839%23issuecomment-560048729>. Please let me know the version number of the archive and the DOI that you receive. Thanks!
