### 2025-03-18 #7323 @jcross186
Great! Here is the DOI:
10.5281/zenodo.15047374

From: Mark Jensen ***@***.***>
Date: Monday, March 17, 2025 at 6:00 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

You can go ahead with the archive @jcross186<https://secure-web.cisco.com/1gJCFnQR2n5nFpCMXXEdlrwpV1XTvcULcCpjIPvOLv1ZJhVVlkHFVtol4khlmXiO82WANbYVhSVoVXfh-P1K7Kko4nAudY6NZDZ4O9PWGXj-lYbS4gusWcdNdqucX_jUmIdg0Yod2s8J4fxhEP0KGneT82ooUZXQKLy-drXTPwcUbofl45l2y4pXGqJ4ar5QLmtf9Js0RvIq9Y4M2azmnaS2o93wDWGUO-SyB1qVxjwN-mtWpvr5AAcJ1UuG-WHFaefiTmBLlv6zvL--XHoDnl8FRR7trRZOENtz0cYFk58Dk_4YQLRuoA37e6eW3vUgm/https%3A%2F%2Fgithub.com%2Fjcross186>
