### 2025-03-20 #7323 @majensen
@jcross186 thanks for the archive -
Can you change the name to the title of the paper? This is a JOSS convention.
