### 2025-03-20 #7323 @majensen
@jcross186 This is a nicely written paper, thank you! Minor copy edits follow.
l. 32 - i.e. Markov -> i.e., Markov
l. 33 - "label switching", ref or brief description?
l. 53 - as -> and
l 87 - missing full stop
l. 132 - ref journal/url?
l. 127 - ref volume?
l. 161 - ref repo/ url?

Picky: parallelize the bullets ll. 46-59 ('models simultaneously', 'uses Bayesian...', ..., 'allows', 'provides')
