### 2025-03-20 feb21bc
Adding DOIs and making minor changes to manuscript to improve readability
