+++ paper.bib
@@ -1,9 +1,12 @@
 @misc{tan2022bcclong,
-  title={BCClong package: Bayesian consensus clustering model for mixed-type longitudinal data},
-  author={Tan, Z and Shen, C and Lu, Z},
-  year={2022}
-}
-
+    title = {A Joint Modeling Approach for Clustering Mixed-Type
+      Multivariate Longitudinal Data: Application to the CHILD Cohort
+      Study},
+    author = {Zhiwen Tan and Zihang Lu and Chang Shen},
+    year = {2022},
+    url = {https://doi.org/10.48550/arXiv.2210.08385},
+    doi={10.32614/CRAN.package.BCClong}
+  }
 
 @article{komarek2014capabilities,
   title={Capabilities of R package mixAK for clustering based on multivariate continuous and discrete longitudinal data},
@@ -11,14 +14,16 @@
   journal={Journal of Statistical Software},
   volume={59},
   pages={1--38},
-  year={2014}
+  year={2014},
+  doi={10.18637/jss.v059.i12}
 }
 
 
 @article{komarek2013clustering,
   title={Clustering for multivariate continuous and discrete longitudinal data},
   author={Kom{\'a}rek, Arno{\v{s}}t and Kom{\'a}rkov{\'a}, Lenka},
-  year={2013}
+  year={2013},
+  doi={10.1214/12-AOAS580}
 }
 
 
@@ -30,7 +35,8 @@
   number={20},
   pages={2610--2616},
   year={2013},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btt425}
 }
 
 
@@ -39,7 +45,8 @@
   author={Lu, Zihang},
   journal={International Statistical Review},
   year={2024},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/insr.12588}
 }
 
 
@@ -51,7 +58,8 @@
   number={29},
   pages={5513--5540},
   year={2023},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/sim.9917}
 }
 
 
@@ -62,7 +70,8 @@
   volume={8},
   number={89},
   pages={5738},
-  year={2023}
+  year={2023},
+  doi={10.21105/joss.05738}
 }
 
 @article{shenk2021traja,
@@ -72,7 +81,8 @@
   volume={6},
   number={63},
   pages={3202},
-  year={2021}
+  year={2021},
+  doi={10.21105/joss.03202}
 }
 
 @article{zang2022bayesian,
@@ -103,7 +113,8 @@
   title={pandas-dev/pandas: Pandas 1.0. 5},
   author={Reback, Jeff and McKinney, Wes and Van Den Bossche, Joris and Augspurger, Tom and Cloud, Phillip and Klein, Adam and Hawkins, Simon and Roeschke, Matthew and Tratner, Jeff and She, Chang and others},
   journal={Zenodo},
-  year={2020}
+  year={2020},
+  doi={10.5281/zenodo.3898987}
 }
 
 
@@ -114,7 +125,8 @@ booktitle = {Advances in Neural Information Processing Systems 32},
 pages = {8024--8035},
 year = {2019},
 publisher = {Curran Associates, Inc.},
-url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
+url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf},
+doi={10.48550/arXiv.1912.01703}
 }
 
[…]
