### 2025-03-20 #7323 @jcross186
HI  @majensen<https://secure-web.cisco.com/1nZacZGLNfIyiYuSXsbboUINsr-HuwM2ASuqywnVNPD4pGprvoZlrlnQU4ZNmOnWigm4iLBbS89RvnHrTfQNx37dlHuBH-zoUEM5MYdcjD6obH8muOEN5OkUd8iaUWPCcjpvrMmygyz8qar5egnq1qFVPWevyhi6jzGG4mbYl3pts0w8xvckSWGjfqaTAZsOx9YoVlj1K5eOSrZilFMoTXCTlla6xPs5xRDCau4l0YCsZORZ-eHwhQpRBdBGADoaU0c24JsN0nSAs7tFT1w5FXczUG9a7ZuWfmUc-ySklYwiIyCkClK7BmxbWBOr1uk0E/https%3A%2F%2Fgithub.com%2Fmajensen> – I have added them. Thanks!

From: Mark Jensen ***@***.***>
Date: Thursday, March 20, 2025 at 12:19 AM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

@jcross186<https://secure-web.cisco.com/1GmGU4F4jPax27qYmFcsV7PUEJVzvpJ9WlJFYlZ2Cio71ObpgvVWcHcrAvPORvPs6J5Y6eI2noZouNMEsJc8AD8jf7oSr_rbxuhg4IePdfDTgQL-RmLInx-oknQe0yJLiR3Ar-oQUmz0iy5-euf4msmc-jNSOHSZ2t6Tf0fyv2lsec5p6U0EO5JUjGarlJ7Z5G6yPjIxpWaYPIlOaW6ZL6dOqJn4eRBuKZDThJ-8RFrG7FUJdk0Q08nR-w-96BxVV8Wd8PKpxCEsJgp0IAeqP2xUmzvCyIJKb_FcIuHNsakdg41OYoEAl2LrMgfd-VCju/https%3A%2F%2Fgithub.com%2Fjcross186> finally, will you add the "missing dois" identified by editorialbot above to your bibtex? Thanks!
