### 2025-03-20 #7323 @jcross186
@majensen<https://secure-web.cisco.com/1nZacZGLNfIyiYuSXsbboUINsr-HuwM2ASuqywnVNPD4pGprvoZlrlnQU4ZNmOnWigm4iLBbS89RvnHrTfQNx37dlHuBH-zoUEM5MYdcjD6obH8muOEN5OkUd8iaUWPCcjpvrMmygyz8qar5egnq1qFVPWevyhi6jzGG4mbYl3pts0w8xvckSWGjfqaTAZsOx9YoVlj1K5eOSrZilFMoTXCTlla6xPs5xRDCau4l0YCsZORZ-eHwhQpRBdBGADoaU0c24JsN0nSAs7tFT1w5FXczUG9a7ZuWfmUc-ySklYwiIyCkClK7BmxbWBOr1uk0E/https%3A%2F%2Fgithub.com%2Fmajensen> – I have made this change. Thank you!

From: Mark Jensen ***@***.***>
Date: Wednesday, March 19, 2025 at 11:42 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

@jcross186<https://secure-web.cisco.com/1JHgzCnOoxNpBlFx4Gfl35KdLQ2hN5hkEDxc7B85g0bJwkXCEyrA-zii6OTRpPdLrqLbwAz1i_fus-vOrqWOu1IfXoCqmh_pvJ7pzRf5dS5xwKDnA4MXNGYAU4OJprHz20wAvTbg5chrbD0dZMYHNtHgv6gGecBZGM12WKyLFDykHjbXXYfsfFL4wFbbmI_X45SOyB33H2HTofBhG6CkuZlEe8I-OuN1_Vp6dYylAICfOnhxs1Lc5K6ocOzDPoFZkrfOIuugQy-DUkiK6SYaZ0_Lfu1N4AUFNgveVvs25L1-Vf80v9BVeNTNsq5foRw4y/https%3A%2F%2Fgithub.com%2Fjcross186> thanks for the archive -
Can you change the name to the title of the paper? This is a JOSS convention.
