### 2025-03-22 #7323 @majensen
Thanks @jcross186 - yes the van der Nest DOI resolves, not sure what happened there, and the Lu reference looks good. I will attempt to summon the Associate Editor-in-Chief.
