### 2025-03-22 #7323 @majensen
@openjournals/dev Can you please have a look at "Error preparing paper acceptance" in https://github.com/openjournals/joss-reviews/issues/7323#issuecomment-2745413896 ? Thanks
