### 2025-03-26 #7323 @arfon
This is what Copilot is suggesting at least...
