### 2025-04-06 #7323 @Kevin-Mattheus-Moerman
👉 Author set does not match, please edit the archive listing so the authors match the paper authors.
👉 No tagged release is found on the GitHub repository. Please create a tagged release and use the tag `1.0.2` (or create a new version and a new archive link to go with it for that version).

👉 Please spell out `USA` as `United States of America`
