### 2025-04-10 ff12d3c
Adding test for update_w_gaussian
