+++ bayes_traj/tests/test_ranefs.py
@@ -300,3 +300,140 @@ def test_update_lambda():
         "Estimate precision variance not as expected"
     assert np.isclose(traj_1_est_var, traj_1_measured_var, rtol=0, atol=1e-5), \
         "Estimate precision variance not as expected"    
+
+def test_update_w_gaussian():
+    """
+    Create synthetic data: 5 individuals with 5 time points each. Two 
+    trajectories, and each individual is associated with random effect
+    vectors drawn from multivariate normal distributions.
+    """
+    np.random.seed(42)
+    n_long_pts = 5
+    D = 1
+    M = 2
+    K = 3
+    G = 5
+    N = G*n_long_pts
+
+    mu1 = np.array([10, -3/5])
+    mu2 = np.array([-10, -3/10])
+    covmat = np.array([[ 0.03      , -0.02999999],
+                       [-0.02999999,  0.03      ]])
+
+    n_dom_pts = 100
+    n_samples1 = 3
+    n_samples2 = 2
+    dom = np.linspace(0, 10, n_dom_pts)
+    samples1 = np.random.multivariate_normal(mean=mu1,
+        cov=covmat, size=n_samples1)
+    samples2 = np.random.multivariate_normal(mean=mu2,
+        cov=covmat, size=n_samples2)
+    sig1 = .2
+    sig2 = .15
+
+    X = np.ones([n_long_pts, 2])
+    X[:, 1] = np.linspace(0, 10, n_long_pts)
+    y = np.zeros(N)
+    for ii in range(n_samples1):
+        y[ii*n_long_pts:(ii+1)*n_long_pts] = \
+            np.dot(X, samples1[ii, :]) + sig1*np.random.randn(n_long_pts)
+
+    for ii in range(n_samples2):
+        y[(ii+n_samples1)*n_long_pts:(ii+n_samples1+1)*n_long_pts] = \
+            np.dot(X, samples2[ii, :]) + sig2*np.random.randn(n_long_pts)
+
+    target_names = ['y']
+    predictor_names = ['intercept', 'x']
+    df = pd.DataFrame()
+    df['intercept'] = np.ones(N)
+    df['id'] = ['a']*n_long_pts + ['b']*n_long_pts + ['c']*n_long_pts + \
+        ['d']*n_long_pts + ['e']*n_long_pts
+    df['x'] = [0, 2.5, 5, 7.5, 10]*G
+    df['y'] = y
+
+    R = np.zeros([N, K])
+    R[0:15, 0] = 1
+    R[15::, 1] = 1
+
+    alpha = 1.242 # For 5 individuals and 2 exptected trajectories
+    
+    w_mu0 = np.zeros([M, D])
+    w_mu0[0, 0] = np.mean(np.array([mu1[0], mu2[0]]))
+    w_mu0[1, 0] = np.mean(np.array([mu1[1], mu2[1]]))
+
+    w_var0 = np.ones([M, D])
+    w_var0[0, 0] = np.var(np.array([mu1[0], mu2[0]]))
+    w_var0[1, 0] = np.var(np.array([mu1[1], mu2[1]]))
+
+    prec_weight = 1e1
+    lambda_a0 = prec_weight*35*np.ones([D])
+    lambda_b0 = prec_weight*np.ones([D])
+    Sig0 = {'y': 1e0*torch.from_numpy(covmat)}
+    ranef_indices = np.ones(M, dtype=bool)
+
+    lambda_a = np.array([[42.5000, 40.0000, 35.0000]])
+    lambda_b = np.array([[1.1327, 1.0798, 1.0000]])
+
+    u_mu = \
+        torch.from_numpy(np.array([[[[-0.04736315,  0.04736312],
+                                     [ 0.        ,  0.        ],
[…]
