+++ bayes_traj/tests/test_mult_dp_regression.py
@@ -738,176 +738,6 @@ def test_augment_df_with_traj_info():
     
     df_aug2 = mm.augment_df_with_traj_info(df)
 
-def test_update_u():
-    """
-    """
-    num_dom_pts = 3
-    sig = 0.2
-    np.random.seed(42)
-    eps = sig*np.random.randn(3*num_dom_pts)
-
-    # Ground truth param values for three "individuals" (columns are intercept
-    # and slope)
-    params = np.zeros((3, 2))
-    x = np.linspace(0, 100, num_dom_pts)
-    params[0, :] = np.array([5, -5/100])
-    params[1, :] = np.array([7, -7/100])
-    params[2, :] = np.array([3, -3/100])
-    y = params[0, 0] + params[0, 1]*x + eps[0:num_dom_pts]
-    y1 = params[1, 0] + params[1, 1]*x + eps[num_dom_pts:2*num_dom_pts]
-    y2 = params[2, 0] + params[2, 1]*x + eps[2*num_dom_pts:3*num_dom_pts]
-        
-    M = 2
-    D = 1
-    w_var0 = np.ones([M, D])
-    w_mu0 = np.zeros([M, D])
-    lambda_a0 = np.ones(D)
-    lambda_b0 = np.ones(D)
-    alpha = 1.
-    K = 10
-    mm = MultDPRegression(w_mu0, w_var0,
-                          lambda_a0, lambda_b0,
-                          1, 1, K=K)
-    N = 3*num_dom_pts
-        
-    X_mat = np.ones((N, M))
-    X_mat[0:num_dom_pts, 1] = x
-    X_mat[num_dom_pts:2*num_dom_pts, 1] = x
-    X_mat[2*num_dom_pts:3*num_dom_pts, 1] = x
-    
-    Y_mat = np.zeros(3*num_dom_pts)
-    Y_mat[0:num_dom_pts] = y
-    Y_mat[num_dom_pts:2*num_dom_pts] = y1
-    Y_mat[2*num_dom_pts:3*num_dom_pts] = y2
-    
-    mm.M_ = M
-    mm.D_ = D
-    mm.X_ = torch.from_numpy(X_mat)
-    mm.Y_ = torch.from_numpy(Y_mat).unsqueeze(-1)
-    mm.predictor_names_ = ['intercept', 'x']
-    mm.target_names_ = ['y']
-    
-    mm.sig_trajs_ = torch.zeros(K, dtype=bool)
-    mm.sig_trajs_[0] = True
-    
-    R = np.zeros([N, K])
-    R[:, 0] = 1
-    mm.R_ = torch.from_numpy(R)
-    
-    w_mu = torch.zeros([M, D, K])
-    w_mu[:, 0, 0] = torch.tensor([params[0, 0], params[0, 1]])
-    mm.w_mu_ = w_mu
-    
-    df = pd.DataFrame()
-    df['x'] = X_mat[:, 1]
-    df['intercept'] = X_mat[:, 0]
-    df['y'] = Y_mat[:]
-    df['id'] = ['a']*num_dom_pts + ['b']*num_dom_pts + ['c']*num_dom_pts
-    
-    mm.df_ = df
-    mm.gb_ = mm.df_.groupby('id')
-    
-    lambda_a = torch.ones([D, K])
-    lambda_b = torch.ones([D, K])
-    lambda_a[0, 0] = 377
-    lambda_b[0, 0] = 45
-    mm.lambda_a_ = lambda_a
-    mm.lambda_b_ = lambda_b
-    
-    group_first_index = np.zeros(N, dtype=bool)
-    group_first_index[0] = True
-    group_first_index[num_dom_pts] = True
-    group_first_index[2*num_dom_pts] = True
[…]
