+++ bayes_traj/mult_dp_regression.py
@@ -1,4 +1,5 @@
 import torch
+from torch.linalg import eigvalsh
 from torch.distributions import Normal, Gamma, Beta, constraints
 from torch.distributions import MultivariateNormal, Bernoulli
 import pyro
@@ -169,7 +170,7 @@ class MultDPRegression:
             if 'Sig0' in kwargs.keys():
                 self.Sig0_ = kwargs['Sig0']
             if 'ranef_indices' in kwargs.keys():
-                self.ranef_indices_ = kwargs['ranef_indices']                
+                self.ranef_indices_ = kwargs['ranef_indices'].astype(bool)
             if 'prob_thresh' in kwargs.keys():
                 self.prob_thresh_ = kwargs['prob_thresh']                
                 
@@ -543,12 +544,12 @@ class MultDPRegression:
         # If ranefs specified, precompute necessary quantities for speed
         #-----------------------------------------------------------------------
         if self.ranef_indices_ is not None:
-            num_ranefs = np.sum(self.ranef_indices_)
+            num_ranefs = int(np.sum(self.ranef_indices_))
             
             # Precompute the outer product of the predictors, which will be reused
             if self.G_ == self.N_:
                 raise RuntimeError("Specified random effects, but no groups")
-    
+
             X_outer = torch.einsum('ij,ik->ijk', self.X_[:, self.ranef_indices_],
                                    self.X_[:, self.ranef_indices_])
             self.G_r_r_ = torch.zeros([self.G_, num_ranefs, num_ranefs])
@@ -565,8 +566,8 @@ class MultDPRegression:
         # inference. If not, it will remain zero.
         self.u_mu_ = torch.zeros((self.G_, self.D_, self.K_, self.M_))
         self.u_Sig_ = \
-            torch.zeros((self.G_, self.D_, self.K_, self.M_, self.M_), \
-                    dtype=torch.float64)
+            1e-20*torch.ones((self.G_, self.D_, self.K_, self.M_, self.M_), \
+                             dtype=torch.float64)
         
         # w_covmat_ is used for binary target variables. The EM algorithm
         # that is used to estimate w_mu_ and w_var_ for binary targets
@@ -614,7 +615,7 @@ class MultDPRegression:
                     self.init_R_mat(traj_probs, traj_probs_weight)
                     if torch.sum(self.sig_trajs_).item() == num_init_trajs:
                         break
-                
+
         self.fit_coordinate_ascent(iters, verbose, weights_only)
 
 
@@ -658,28 +659,25 @@ class MultDPRegression:
         inc = 0
         while inc < iters:
             inc += 1
-
+            
             self.update_v()
+            if (self.ranef_indices_ is not None):
+                if np.sum(self.ranef_indices_) > 0:
+                    self.update_u()            
             if self.num_binary_targets_ > 0:
                 self.update_w_logistic(em_iters=1)
             if (self.D_ - self.num_binary_targets_ > 0) and \
                (not weights_only):
                 self.update_w_gaussian()
-                self.update_lambda() 
-
+                self.update_lambda()
             self.R_ = self.update_z(self.X_, self.Y_)
 
-            if self.ranef_indices_ is not None:
-                if np.sum(self.ranef_indices_) > 0:
-                    self.update_u()
-            
-            self.sig_trajs_ = \
-                torch.max(self.R_, dim=0).values > self.prob_thresh_
+            self.sig_trajs_ = torch.max(self.R_, dim=0).values > self.prob_thresh_
 
             if verbose:
                 torch.set_printoptions(precision=2)
-                #print(self.w_mu_[:, 0, 0])
[…]
