### 2025-04-10 e7cdfe5
Adding test for update_z
