### 2025-04-10 #7323 @jcross186
I have addressed the highlighted issues. Please confirm and let me know if anything else is needed.

Kind regards,

-James

From: Kevin Mattheus Moerman ***@***.***>
Date: Sunday, April 6, 2025 at 2:10 AM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

@jcross186<https://secure-web.cisco.com/1W4dGyWkcZpoeVuxZ-s6dBFnkCYxgD5FspMHZAb1vMpgWGgwnaG41YqH-h-zrKn-wwe3SOLygGDrGNn_vHudpuo3k8sp0rjlheraTZ2Gy2NOd09Jy9S3UYnFtKOPR_YrsxcWFhwcee8yU0DqKSlaoMa5Ho_fQU1KN48YV5szPyKCa2hg5l4yu_KfD9y5HiZi2zsiFfwOavJhBJ6Z7rMk1tEVA3y07SSkcgp79oX3oPTDQyy0_zdp4N3qVx-8YzzGofyt-yBWlTIAohptHxXiCgt3-6vqxqdmFaJIt7VwZooSs-CDHjYLmg6EG2-8HoYYt/https%3A%2F%2Fgithub.com%2Fjcross186> as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:

Checks on repository

  *   Project has OSI approved license
  *   Project features contributing guidelines

Checks on review issue

  *   Review completed
  *   Software version tag listed here matches a tagged release

Checks on archive

  *   Archive listed title and authors matches paper
👉 Author set does not match, please edit the archive listing so the authors match the paper authors.
  *   Archive listed license matches software license
  *   Archive listed version tag matches tagged release (and includes a potential v).
👉 No tagged release is found on the GitHub repository. Please create a tagged release and use the tag 1.0.2 (or create a new version and a new archive link to go with it for that version).

Checks on paper

  *   Checked paper formatting
  *   Check affiliations to make sure country acronyms are not used
👉 Please spell out USA as United States of America
  *   Checked reference rendering
  *   Checked if pre-print citations can be updated by published versions
  *   Checked for typos
