### 2025-04-14 #7323 @jcross186
I’m seeing ‘United States of America’ in both my and Dr. Zhao’s affiliation. Was there something else that needed to be changed?

Thanks,

-James

From: Kevin Mattheus Moerman ***@***.***>
Date: Monday, April 14, 2025 at 8:42 AM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

@jcross186<https://secure-web.cisco.com/1k48OYBuuvvf3iya3DGf1xYMQf7SsJR5ZMbzI9HhJeSv8j_GJaTDZjxfYjfX8sLsnvPpThgqYK8YsNO45Qm0u2BCuHqHQw3BTiE2NIy9bvEdrtS8m_-13zKUS1vglHE7osoMnJIkAHwglk-H4gp2OjH4cwYYip7j1aAN2zDwcVng5iI7KeNHadJEbkKOlSMK9vOI-xG2Pecng8ZBCahCIiUILOnH9oC6sX0TimvpTzI1oGxb4SqFhyPm3T5gJ_2CabwSCLik5c1iiFAPT6uBB6ap0oHfqDiUzRsOdDj4_dEgS_XYNqtVKotumMim-l0Uv/https%3A%2F%2Fgithub.com%2Fjcross186> can you check the affiliation update? I cannot see this reflected in the paper yet.
