### 2025-04-14 336839c
Gets K from prior file if possible
