### 2025-04-14 adbb431
DEB: fixed way in which interactions are handled
