### 2025-04-15 #7323 @Kevin-Mattheus-Moerman
@jcross186 the latest draft shows `USA` which should read `United States of America`:

![Image](https://github.com/user-attachments/assets/9810ca6d-30db-4016-b7bf-4661a07c6828)

and so does the `paper.md` source on the `joss` branch:

![Image](https://github.com/user-attachments/assets/8bbc9e19-03e2-47d4-b336-467b6f4338de)
