### 2025-04-15 #7323 @jcross186
Should be updated now…

From: Kevin Mattheus Moerman ***@***.***>
Date: Tuesday, April 15, 2025 at 5:42 AM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

@jcross186<https://secure-web.cisco.com/1kVlADda1pN3GDJIObWPMXJwG4LjKHYBGbL0AdHI-lysKgrs_b-I6tExGrCgs9tpxdnq-BjetqdKqGMD6269X3HbJoVDhDUGRJKhF6Z8y2H9J-106N3CdR-PT6a6QlHjMQjqmQyQdl-awP258oLmcGccoKJ_x2o6j8kKiFGL2J44fv8hxxlv6vpk53QWowgp2BF74qjnvSEpw6C6Whx2Nz6Qfz61dCtr1lIE81dPMBzyDcMhRDSqH1YpxraI-yUkn22IrOfVVr0FWt_QlzJYgTB3Kr113jhwTPqIGRf7JPsUjR3GQ1feeAX4vpBD9JEkB/https%3A%2F%2Fgithub.com%2Fjcross186> the latest draft shows USA which should read United States of America:

image.png (view on web)<https://secure-web.cisco.com/1rbo0fLKbzOqnq77PFnwS1UzSwiNAcLdb1HcH9XvOTavRWEv_R_WTYsSXwtFexhv0QdIyi989YxzAr-Qxmn5LQ4EOoZN-072EkYQj3fNZc0prv9no1_g_ydvyRpvZAtB1Zw5hnp2cmWbNmhnEMcEgILJsgRBcfBhrEaBgQWeWXrSAIEYeix302JBQkLCb-O9DqXcelp3u-CSiiONgjVcC2PVTYIoC9wmzrqiOzkpBSuk5FLp0Li2UgqXjVUB6_e2rt2KjHeBtVG4CsbFfef_deQl4shlHL3l6COUGDNGtWpA8CpQib75JeNdNyB_9-KKD/https%3A%2F%2Fgithub.com%2Fuser-attachments%2Fassets%2F9810ca6d-30db-4016-b7bf-4661a07c6828>

and so does the paper.md source on the joss branch:

image.png (view on web)<https://secure-web.cisco.com/10-X9hRQw4xS4W8MOUozHIev9w3sAisLjJKmJaBg_RH4apV7W6xVUwUCUPlF0dy57eQB85cupojbFisO1oJw2jA_ngJgvtIp0ntq-1CVM0ei714mGzeT01J-H48k-1KIyEzR-2LMB49Pj-QYsAu2F9bIoY5UHQXuFBesPFWbpOATEmAV4OC2JZaDC807QCKzj_IXkjG2Xqa5gzStdl1IL71f5UR9f7HamgSQto2ceGs3P_eHAFC6ejkHHNZrdci9ZghuspqQsTbeEWO_SxIbbetqQzsqEEaEoYYQCNRup9qNaahhlvEHp15UzlU96HTba/https%3A%2F%2Fgithub.com%2Fuser-attachments%2Fassets%2F8bbc9e19-03e2-47d4-b336-467b6f4338de>
