### 2025-04-16 #7323 @jcross186
Yes, thanks to all for your time and effort! Much appreciated!

-James

From: Kevin Mattheus Moerman ***@***.***>
Date: Tuesday, April 15, 2025 at 5:17 PM
Cc: Ross, James Chapman,Ph.D. ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: bayes_traj: A Python package for Bayesian trajectory analysis (Issue #7323)

        External Email - Use Caution

@jcross186<https://secure-web.cisco.com/10U0mUAAxN_zJucHESdiCvMdTBmdNlraUd4NzGxke75Kp-iOcfWMw_e0dL2LqAkXpqF4XBxqZy6mQH2mcSAhSVL_taXie-sh4JaZftMSAwyeuHOkki7Srd7RI5mCMaRLOKL4zDxvK7Mh1z23OrspondjZp6wNn6Jrt2GhgfMSgq2cZ1rsXZ04WjbWAvrLPvHiD1JLVe-yvsc2kwUehzrO7ixlB-KvlfGOV9kWUjAliV0AeLOPViDCt3zuWFISZJ37hzgkjyRMPMRirZayEQc6tXkN2mKJ3mjABmOjyxd2qwPz0-t69pKQoZekQm8jswja/https%3A%2F%2Fgithub.com%2Fjcross186> congratulations on this JOSS publication!

Thanks for editing @majensen<https://secure-web.cisco.com/1PXK5Iv-KRLO94REbPvNEItyalc7ExwF1Jh00VgYhoDg-vHgnQVaOE1OmPu3cn_fqRPfQnF6WNahxthSgSf1tWDcd0l-5gKlfnQ53LuQsu4H0UB-4eG1TzIDQ4yWqoMwPgXDCdDD1iJozt8WrUssZ81kCC-w2cjtIQSDpS3Ka9l4d2WMgKdWdaHlD3nwuznOll9PpLRviMaz6IWTPsVoUZvqKooJKV2JVdY9MqbdPD-7mlP7bb5uh4IG99VGrl6aZXF_K85N1a_T77iZmJ14JREH9zK4956svXqlAGAS4nYHeRNHLpSZ4glvsFHHj9xAx/https%3A%2F%2Fgithub.com%2Fmajensen> !!

And a special thank you to the reviewers: @gvieralopez<https://secure-web.cisco.com/1DBTeXI__bw3bOUjFJMemPiiZoetEjEt7_Hjdtv6pKAbcgrTOMxKfka4N-aoMoS-cTbxKYVvJuKQBHwyMsqn5TKQ7rHppwDe35f-k61FB6_Hjvh6Pi8Ja3i2h652E9dP4GuxKeos2bv3B4YLAu-Y36EXSveW7tyMd4W6Rqk3MCU2fomqvV1eclr9Saadolq4H3ylvRxpjvos1ovGWp1bsylJZkqc52voCY8ZfY9P16sSdjExShPEPNxQMgDAQJRXoASOBxxocYFRbqWjqlBNZqDeceYV6AvOhCLK4FovfeY6-w0hJ-sYsKoEzojA4hz_h/https%3A%2F%2Fgithub.com%2Fgvieralopez>, @gchure<https://secure-web.cisco.com/106fpPeFwO5KVRzU3dqJfTwxKoGC3S8iaEKpWbMrqGQ86WeVmdhXyQMpn2fLnqj_bbDM9l_vExP_6p0gYr73eFPQMk5VYLhKy8Chh2QwolHedNDs5xmO5kYvohHcBwZDc9fuACGHSsWXWkAxy_xHONCLJA2E8zvPkJKjWt1LkvmlULsSVJlk66r3ooiYsbcLM4KQhWaFnbvqfj_4byCbiQyihdJB44Np_Y5O2oBadGDpWXtjb0hC4vE3deyirl768BdkYixC61rAxyM2T1000jaDqAPq-dZBS6V5VST1M08M0p4Sk8b1yycwAtLZOKlxL/https%3A%2F%2Fgithub.com%2Fgchure> !!
